Add AuthorizationCheckActivity
Intergrate authority check function in the situation of  encounting editing an Item.
If user wants to edit an item, gesture check shall be verified firstly.

By Using startActitivityForResult method, PasswordListAdapter which detect the click event on Item's "Edit" option  will initial AuthorizationCheckActivity firstly. The check process will be handled in AuthorizationCheckActivity. 
If the gesture check is not correct, "Incorrect Password" information will be reveal on the interface.
Once the check was clear to go, the result which contains an messange(RESULT_OK) and the Intent(contains the information for EditAcitivity) will be sent to the Activity that contains the PasswordListAdapter.
MainActivity, in this project, contains PasswordListAdapter, will take care of the result to grant the initial of EditAcitivity.

In all and all, AuthorizationCheckActivity acts as the intermidiate which verfifys the gesture code to decide whether to initial EditAcitivity or not.